# JobsAPI/swagger.yaml
openapi: 3.0.0
info:
  title: JobsAPI
  contact: {}
  version: '1.0'
servers:
- url: http://localhost:3000/api/v1
  variables: {}
paths:
...
  /jobs:
    post:
      tags:
      - Jobs
      summary: createJobs
      operationId: createJobs
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createJobsRequest'
            example:
              company: Company Name
              position: Postion
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    get:
      tags:
      - Jobs
      summary: getAllJobs
      operationId: getAllJobs
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
...
    createJobsRequest:
      title: createJobsRequest
      required:
      - company
      - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      example:
        company: Company Name
        position: Position Name
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...